feat: optimize MissionFactory gas usage
Changed `missionCount` to `uint96` to pack it with `disputeResolver` in the same storage slot.
This saves ~20k gas on the first mission creation and ~2100 gas on every subsequent mission creation.
Verified with `forge test`.

+++ src/MissionFactory.sol
@@ -1,13 +1,13 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.24;
 
-import {Clones} from "@openzeppelin/contracts/proxy/Clones.sol";
-import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import {SafeERC20} from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
-import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
-import {ReentrancyGuard} from "@openzeppelin/contracts/utils/ReentrancyGuard.sol";
-import {IMissionEscrow} from "./interfaces/IMissionEscrow.sol";
-import {MissionEscrow} from "./MissionEscrow.sol";
+import { Clones } from "@openzeppelin/contracts/proxy/Clones.sol";
+import { IERC20 } from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
+import { SafeERC20 } from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
+import { Ownable } from "@openzeppelin/contracts/access/Ownable.sol";
+import { ReentrancyGuard } from "@openzeppelin/contracts/utils/ReentrancyGuard.sol";
+import { IMissionEscrow } from "./interfaces/IMissionEscrow.sol";
+import { MissionEscrow } from "./MissionEscrow.sol";
 
 /**
  * @title MissionFactory
@@ -38,7 +38,8 @@ contract MissionFactory is Ownable, ReentrancyGuard {
     address public disputeResolver;
 
     /// @notice Current mission counter
-    uint256 public missionCount;
+    /// @dev Packed with disputeResolver (20 bytes + 12 bytes = 32 bytes)
+    uint96 public missionCount;
 
     /// @notice Mapping from mission ID to escrow address
     mapping(uint256 => address) public missions;
@@ -93,10 +94,7 @@ contract MissionFactory is Ownable, ReentrancyGuard {
      * @param _usdc USDC token address
      * @param _paymentRouter PaymentRouter contract address
      */
-    constructor(
-        address _usdc,
-        address _paymentRouter
-    ) Ownable(msg.sender) {
+    constructor(address _usdc, address _paymentRouter) Ownable(msg.sender) {
         usdc = IERC20(_usdc);
         paymentRouter = _paymentRouter;
 
@@ -145,18 +143,19 @@ contract MissionFactory is Ownable, ReentrancyGuard {
         address escrow = escrowImplementation.clone();
 
         // Initialize escrow
-        IMissionEscrow(escrow).initialize(
-            missionId,
-            msg.sender,
-            rewardAmount,
-            expiresAt,
-            guild,
-            metadataHash,
-            locationHash,
-            paymentRouter,
-            address(usdc),
-            disputeResolver
-        );
+        IMissionEscrow(escrow)
+            .initialize(
+                missionId,
+                msg.sender,
+                rewardAmount,
+                expiresAt,
+                guild,
+                metadataHash,
+                locationHash,
+                paymentRouter,
+                address(usdc),
+                disputeResolver
+            );
 
         // Store mission mapping
         missions[missionId] = escrow;
@@ -195,10 +194,10 @@ contract MissionFactory is Ownable, ReentrancyGuard {
      * @param missionId The mission ID
      * @return params The mission parameters
      */
-    function getMissionParams(uint256 missionId) 
-        external 
-        view 
-        returns (IMissionEscrow.MissionParams memory params) 
+    function getMissionParams(uint256 missionId)
+        external
+        view
+        returns (IMissionEscrow.MissionParams memory params)
     {
         address escrow = missions[missionId];
         if (escrow == address(0)) revert MissionNotFound();
@@ -210,10 +209,10 @@ contract MissionFactory is Ownable, ReentrancyGuard {
      * @param missionId The mission ID
      * @return runtime The mission runtime state
      */
-    function getMissionRuntime(uint256 missionId) 
-        external 
-        view 
-        returns (IMissionEscrow.MissionRuntime memory runtime) 
+    function getMissionRuntime(uint256 missionId)
+        external
+        view
+        returns (IMissionEscrow.MissionRuntime memory runtime)
     {
         address escrow = missions[missionId];
         if (escrow == address(0)) revert MissionNotFound();
